instavibe/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.90.0
Flask
google-cloud-spanner
humanize